Restructured Model Class and More Callbacks Build
I cleaned up the Model class so that it's more automated.  Added utility functions to serialize and unserialize a stored Dataframe; added similar functions for the Model class.

Results page is early in the works but will have something for the presentation. Easier to work with now that I have a working Model class.